pref: 日志模块ES配置加载与连接优化 #3414 优化退避时长
