+++ src/backend/ci/core/common/common-log/src/main/kotlin/com/tencent/devops/common/log/pojo/event/LogBatchEvent.kt
@@ -24,10 +24,11 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.common.log.pojo
+package com.tencent.devops.common.log.pojo.event
 
 import com.tencent.devops.common.event.annotation.Event
 import com.tencent.devops.common.event.dispatcher.pipeline.mq.MQ
+import com.tencent.devops.common.log.pojo.ILogEvent
 import com.tencent.devops.common.log.pojo.message.LogMessageWithLineNo
 
 /**
@@ -38,6 +39,6 @@ import com.tencent.devops.common.log.pojo.message.LogMessageWithLineNo
 data class LogBatchEvent(
     override val buildId: String,
     val logs: List<LogMessageWithLineNo>,
-    override val retryTime: Int = 2,
+    override val retryTime: Int = 0,
     override val delayMills: Int = 0
 ) : ILogEvent(buildId, retryTime, delayMills)
