+++ src/backend/ci/core/common/common-log/src/main/kotlin/com/tencent/devops/common/log/pojo/event/LogEvent.kt
@@ -24,23 +24,21 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.common.log.pojo
+package com.tencent.devops.common.log.pojo.event
 
 import com.tencent.devops.common.event.annotation.Event
 import com.tencent.devops.common.event.dispatcher.pipeline.mq.MQ
+import com.tencent.devops.common.log.pojo.ILogEvent
+import com.tencent.devops.common.log.pojo.message.LogMessage
 
 /**
  * deng
  * 2019-01-23
  */
-@Event(MQ.EXCHANGE_LOG_STATUS_BUILD_EVENT, MQ.ROUTE_LOG_STATUS_BUILD_EVENT)
-data class LogStatusEvent(
+@Event(MQ.EXCHANGE_LOG_BUILD_EVENT, MQ.ROUTE_LOG_BUILD_EVENT)
+data class LogEvent(
     override val buildId: String,
-    val finished: Boolean,
-    val tag: String,
-    val subTag: String?,
-    val jobId: String,
-    val executeCount: Int?,
-    override val retryTime: Int = 2,
+    val logs: List<LogMessage>,
+    override val retryTime: Int = 0,
     override val delayMills: Int = 0
 ) : ILogEvent(buildId, retryTime, delayMills)
