+++ src/backend/ci/core/common/common-log/src/main/kotlin/com/tencent/devops/common/log/pojo/event/LogStatusEvent.kt
@@ -24,20 +24,24 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.common.log.pojo
+package com.tencent.devops.common.log.pojo.event
 
 import com.tencent.devops.common.event.annotation.Event
 import com.tencent.devops.common.event.dispatcher.pipeline.mq.MQ
-import com.tencent.devops.common.log.pojo.message.LogMessage
+import com.tencent.devops.common.log.pojo.ILogEvent
 
 /**
  * deng
  * 2019-01-23
  */
-@Event(MQ.EXCHANGE_LOG_BUILD_EVENT, MQ.ROUTE_LOG_BUILD_EVENT)
-data class LogEvent(
+@Event(MQ.EXCHANGE_LOG_STATUS_BUILD_EVENT, MQ.ROUTE_LOG_STATUS_BUILD_EVENT)
+data class LogStatusEvent(
     override val buildId: String,
-    val logs: List<LogMessage>,
-    override val retryTime: Int = 2,
+    val finished: Boolean,
+    val tag: String,
+    val subTag: String?,
+    val jobId: String,
+    val executeCount: Int?,
+    override val retryTime: Int = 0,
     override val delayMills: Int = 0
 ) : ILogEvent(buildId, retryTime, delayMills)
