+++ src/backend/ci/core/log/biz-log/src/main/kotlin/com/tencent/devops/log/mq/LogListener.kt
@@ -26,9 +26,9 @@
 
 package com.tencent.devops.log.mq
 
-import com.tencent.devops.common.log.pojo.LogBatchEvent
-import com.tencent.devops.common.log.pojo.LogEvent
-import com.tencent.devops.common.log.pojo.LogStatusEvent
+import com.tencent.devops.common.log.pojo.event.LogBatchEvent
+import com.tencent.devops.common.log.pojo.event.LogEvent
+import com.tencent.devops.common.log.pojo.event.LogStatusEvent
 import com.tencent.devops.log.service.LogService
 import com.tencent.devops.common.log.utils.LogMQEventDispatcher
 import org.slf4j.LoggerFactory
@@ -49,14 +49,14 @@ class LogListener @Autowired constructor(
         } catch (ignored: Throwable) {
             logger.warn("Fail to add the log event [${event.buildId}|${event.retryTime}]", ignored)
         } finally {
-            if (!result && event.retryTime >= 0) {
+            if (!result && event.retryTime <= 2) {
                 logger.warn("Retry to add the log event [${event.buildId}|${event.retryTime}]")
                 with(event) {
                     logMQEventDispatcher.dispatch(LogEvent(
                         buildId = buildId,
                         logs = logs,
-                        retryTime = retryTime - 1,
-                        delayMills = DELAY_DURATION_MILLS * retryTime
+                        retryTime = retryTime + 1,
+                        delayMills = getNextDelayMills(retryTime)
                     ))
                 }
             }
@@ -71,14 +71,14 @@ class LogListener @Autowired constructor(
         } catch (ignored: Throwable) {
             logger.warn("Fail to add the log batch event [${event.buildId}|${event.retryTime}]", ignored)
         } finally {
-            if (!result && event.retryTime >= 0) {
+            if (!result && event.retryTime <= 2) {
                 logger.warn("Retry to add log batch event [${event.buildId}|${event.retryTime}]")
                 with(event) {
                     logMQEventDispatcher.dispatch(LogBatchEvent(
                         buildId = buildId,
                         logs = logs,
-                        retryTime = retryTime - 1,
-                        delayMills = DELAY_DURATION_MILLS * retryTime
+                        retryTime = retryTime + 1,
+                        delayMills = getNextDelayMills(retryTime)
                     ))
                 }
             }
@@ -93,7 +93,7 @@ class LogListener @Autowired constructor(
         } catch (ignored: Throwable) {
             logger.warn("Fail to add the multi lines [${event.buildId}|${event.retryTime}]", ignored)
         } finally {
-            if (!result && event.retryTime >= 0) {
+            if (!result && event.retryTime <= 2) {
                 logger.warn("Retry to add the multi lines [${event.buildId}|${event.retryTime}]")
                 with(event) {
                     logMQEventDispatcher.dispatch(
@@ -104,13 +104,17 @@ class LogListener @Autowired constructor(
                             subTag = subTag,
                             jobId = jobId,
                             executeCount = executeCount,
-                            retryTime = retryTime - 1,
-                            delayMills = DELAY_DURATION_MILLS * retryTime
+                            retryTime = retryTime + 1,
+                            delayMills = getNextDelayMills(retryTime)
                         )
                     )
                 }
             }
         }
+    }
+
+    private fun getNextDelayMills(retryTime: Int): Int {
+        return DELAY_DURATION_MILLS * retryTime
     }
 
     companion object {
